+++ html/bin/console.php
@@ -2,14 +2,14 @@
 
 declare(strict_types=1);
 
-use Otus\DataMapper\Command\DeleteCommand;
-use Otus\DataMapper\Command\EagerCommand;
-use Otus\DataMapper\Command\InsertCommand;
-use Otus\DataMapper\Command\LazyCommand;
-use Otus\DataMapper\Command\RunCommand;
-use Otus\DataMapper\Command\SqlCommand;
-use Otus\DataMapper\Command\UpdateCommand;
 use Otus\DataMapper\Kernel\Console;
+use Otus\DataMapper\Sql\Command\DeleteCommand;
+use Otus\DataMapper\Sql\Command\EagerCommand;
+use Otus\DataMapper\Sql\Command\InsertCommand;
+use Otus\DataMapper\Sql\Command\LazyCommand;
+use Otus\DataMapper\Sql\Command\RunCommand;
+use Otus\DataMapper\Sql\Command\SqlCommand;
+use Otus\DataMapper\Sql\Command\UpdateCommand;
 
 require_once __DIR__ . '/../vendor/autoload.php';
